backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-base64field==1.0
django-colorfield==0.7.2
django-cors-headers==4.1.0
django-filter==21.1
djangorestframework==3.14.0
drf-extra-fields==3.4.0
drf-yasg==1.21.6
fpdf==1.7.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pybase64==1.2.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
